Use all versions for collection, not only preferred
In the situations when there are multiple resources of different
versions in the same group, collection simply skipped older versions
then preferred.

This is fine for the case of outdated or non-served API version, but it
may also appear in scenario when entirely differen CRD is using same
group but is behind in the version. For this scenario, collection needs
to be exhaustive.

Signed-off-by: Danil-Grigorev <danil.grigorev@suse.com>