Simplify local part of release process
Pushing a tag locally would then require a push to the main branch,
which is protected, hence a pull request would be necessary.

In turn, this means that any new version tag would point to a commit on
a feature branch, which does not feel right.

Instead, the release process could run in two steps:
* a local step, from a feature branch: `make update-crate
  version=vX.Y.Z` for updating Cargo files, including a push
* pushing a `vX.Y.Z` tag, which would trigger the release CI workflow.

The next step of this development will be adding a check to the release
workflow to validate consistency between the version specified in Cargo
files and the tag against which the workflow runs.